Create convert_text.c
Converts a text file into an easily readable txt file for the markov babbler. It takes off punctuation, new lines, etc.

I tried to implement this within the babbler, but it caused more errors than results. Maybe with more tinkering this will happen. For now, it is a standalone procedure.